Move expectation types to shared file.
